refactor(persistence-sqlite): rename sqlite_data_store.rs to sqlite_store.rs; drop SqliteDataStore alias
